requirements.txt
aiohttp==3.7.4
alpaca-trade-api==1.5.0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
deprecation==2.1.0
idna==3.3
msgpack==1.0.2
multidict==6.0.2
numpy==1.22.2
packaging==21.3
pandas==1.4.1
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
typing-extensions==4.1.1
urllib3==1.26.8
websocket-client==1.3.1
websockets==9.1
yarl==1.7.2
